[ Bookclub ] 스터디 참가신청/취소

+++ src/main/java/com/example/bookclub/application/StudyService.java
@@ -1,13 +1,18 @@
 package com.example.bookclub.application;
 
+import com.example.bookclub.domain.Account;
+import com.example.bookclub.domain.AccountRepository;
 import com.example.bookclub.domain.Study;
 import com.example.bookclub.domain.StudyRepository;
 import com.example.bookclub.dto.StudyCreateDto;
 import com.example.bookclub.dto.StudyResultDto;
 import com.example.bookclub.dto.StudyUpdateDto;
+import com.example.bookclub.errors.AccountNotFoundException;
 import com.example.bookclub.errors.StartAndEndDateNotValidException;
 import com.example.bookclub.errors.StartAndEndTimeNotValidException;
 import com.example.bookclub.errors.StudyNotFoundException;
+import com.example.bookclub.errors.StudySizeFullException;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -21,12 +26,21 @@ import java.util.List;
 @Transactional
 public class StudyService {
     private final StudyRepository studyRepository;
+    private final AccountRepository accountRepository;
 
-    public StudyService(StudyRepository studyRepository) {
+    public StudyService(StudyRepository studyRepository,
+                        AccountRepository accountRepository) {
         this.studyRepository = studyRepository;
+        this.accountRepository = accountRepository;
     }
 
-    public StudyResultDto createStudy(StudyCreateDto studyCreateDto) throws ParseException {
+    public StudyResultDto createStudy(Account account,
+                                      StudyCreateDto studyCreateDto)
+            throws ParseException {
+        if (account == null) {
+            throw new AccessDeniedException("권한이 없습니다");
+        }
+
         LocalDate startDate = studyCreateDto.getStartDate();
         LocalDate endDate = studyCreateDto.getEndDate();
         LocalDate todayDate = LocalDate.now();
@@ -43,26 +57,62 @@ public class StudyService {
 
         Study study = studyCreateDto.toEntity();
         Study createdStudy = studyRepository.save(study);
+        createdStudy.addAdmin(account.getEmail());
+
         return StudyResultDto.of(createdStudy);
     }
 
-    public StudyResultDto updateStudy(Long id, StudyUpdateDto studyUpdateDto) {
+    public StudyResultDto updateStudy(Account account,
+                                      Long id,
+                                      StudyUpdateDto studyUpdateDto) {
         Study study = getStudy(id);
+        if (!study.isManagedBy(account)) {
+           throw new AccessDeniedException("권한이 없습니다");
+        }
 
         study.updateWith(studyUpdateDto);
-
         return StudyResultDto.of(study);
     }
 
-    public StudyResultDto deleteStudy(Long id) {
+    public StudyResultDto deleteStudy(Account account, Long id) {
         Study study = getStudy(id);
+        if(!study.isManagedBy(account)) {
+            throw new AccessDeniedException("권한이 없습니다");
+        }
+
         studyRepository.delete(study);
         return StudyResultDto.of(study);
+    }
+
+    public Long applyStudy(Account account, Long id) {
+        Study study = getStudy(id);
+        if (study.isSizeFull()){
+            throw new StudySizeFullException();
+        }
+        Account user = getAccount(account.getId());
+
+        study.addAccount(user);
+
+        return id;
+    }
+
+    public Long cancelStudy(Account account, Long id) {
+        Study study = getStudy(id);
+        Account user = getAccount(account.getId());
+
+        study.cancelAccount(user);
+
+        return id;
     }
 
     public Study getStudy(Long id) {
         return studyRepository.findById(id)
                 .orElseThrow(() -> new StudyNotFoundException(id));
+    }
+
+    public Account getAccount(Long id) {
+        return accountRepository.findById(id)
+                .orElseThrow(() -> new AccountNotFoundException(id));
     }
 
     public List<Study> getStudies() {
